==> 3. Demand projection.ipynb <==
# per capita demand projection
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from sklearn.preprocessing import PolynomialFeatures
from sklearn.linear_model import LinearRegression
from sklearn.pipeline import make_pipeline

file_path = 'International Energy Agency - Electricity consumption per capita, Indonesia.csv'
data = pd.read_csv(file_path)

data['Electricity consumption per capita, Indonesia'] = data['Electricity consumption per capita, Indonesia'].astype(float)

X = data[['Year']].values
y = data['Electricity consumption per capita, Indonesia'].values

# Polynomial Regression (Degree 2)
degree = 2
poly_model = make_pipeline(PolynomialFeatures(degree), LinearRegression())
poly_model.fit(X, y)

extended_years = np.arange(data['Year'].min(), 2051).reshape(-1, 1)
extended_predictions = poly_model.predict(extended_years)

year_to_predict = np.array([[2040]])
predicted_consumption_poly = poly_model.predict(year_to_predict)[0]

plt.figure(figsize=(10, 6))

plt.scatter(X, y, color='#3882a4',s=70, label='Actual Data')

plt.plot(extended_years, extended_predictions, color='#f3bc42', label=f'Polynomial Fit (Degree {degree})')

plt.scatter(year_to_predict, predicted_consumption_poly, color='#b74f00',s=70, label='2040 Prediction')

plt.xlabel('Year')
plt.ylabel('Electricity Consumption per Capita (MWh / Capita)')
plt.title('Electricity Consumption per Capita Prediction (Polynomial Regression)')
plt.legend()
plt.grid(True)
plt.show()

print(f"Polynomial Regression Prediction for 2040: {predicted_consumption_poly:.3f} MWh/Capita")
