==> 7. NPV and LCOE.ipynb <==
# NPV
def calc_npv(annual_revenue, i=0.05, lifetime_yrs=25, CAPEX=0, OPEX=0):
    """
    Calculate Net Present Value (NPV)

    Parameters:
    - annual_revenue: Annual revenue
    - i: Discount rate (default = 5%)
    - lifetime_yrs: Project lifespan in years
    - CAPEX: Initial capital expenditure
    - OPEX: Annual operating costs (default = 0)

    Returns:
    - NPV: Net present value, rounded to the nearest integer
    """
    costs_op = [OPEX] * lifetime_yrs
    revenue = [annual_revenue] * lifetime_yrs
    
    t = range(1, lifetime_yrs + 1)  # Year sequence: 1, 2, ..., lifetime_yrs
    npv = sum((revenue[t-1] - costs_op[t-1]) / (1 + i) ** t for t in t) - CAPEX
    
    return round(npv, 0)


annual_revenue = 31780.66
lifetime_yrs = 25
CAPEX = 124393.11
OPEX = 0
discount_rate = 0.05

npv = calc_npv(annual_revenue, i=discount_rate, lifetime_yrs=lifetime_yrs, CAPEX=CAPEX, OPEX=OPEX)

decision = "Support" if npv > 0 else "Reject"
print(f"NPV: {npv}, Decision: {decision}")

# LCOE
def life_span_generation_kwh(yearly_generation_kwh, discount=0.03, lifetime_yrs=25):
    """
    Calculate total discounted lifetime electricity generation.
    
    Parameters:
    - yearly_generation_kwh: Annual electricity generation (in kWh)
    - discount: Discount rate (default = 3%)
    - lifetime_yrs: Project lifespan in years (default = 25)

    Returns:
    - Discounted lifetime electricity generation (in kWh), rounded to the nearest integer
    """
    t = range(1, lifetime_yrs + 1)  # Year sequence: 1, 2, ..., lifetime_yrs
    l_s_g = sum(yearly_generation_kwh / (1 + discount) ** year for year in t)
    return round(l_s_g, 0)


def lcoe(npv_cost, life_span_generation):
    """
    Calculate the Levelized Cost of Energy (LCOE).
    
    Parameters:
    - npv_cost: Net present value (NPV) of total costs (in USD)
    - life_span_generation: Total lifetime electricity generation (in kWh)

    Returns:
    - LCOE (in USD/kWh), rounded to two decimal places
    """
    lcoe_value = npv_cost / life_span_generation
    return round(lcoe_value, 2)


annual_generation =  308550082980  # Annual electricity generation (in kWh)
discount_rate = 0.03
lifetime_years = 25
capex = 124393.11 * 10**6  # Initial capital expenditure (in USD)

lsg = life_span_generation_kwh(yearly_generation_kwh=annual_generation, discount=discount_rate, lifetime_yrs=lifetime_years)
print(f"Discounted Lifetime Electricity Generation (kWh): {lsg}")

# Calculate LCOE
lcoe_value = lcoe(npv_cost=capex, life_span_generation=lsg)
print(f"LCOE (USD/kWh): {lcoe_value}")
